+++ avni-server-api/src/main/java/org/avni/server/service/MetadataDiffChecker.java
@@ -1,0 +1,142 @@
+package org.avni.server.service;
+
+import org.springframework.stereotype.Service;
+
+import java.util.*;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+@Service
+public class MetadataDiffChecker {
+
+    MetadataDiffOutputGenerator metadataDiffOutputGenerator;
+    protected Map<String, Object> findDifferences(Map<String, Object> jsonMap1, Map<String, Object> jsonMap2) {
+        Map<String, Object> differences = new HashMap<>();
+        boolean hasDifferences = false;
+        String uuid = "null";
+        for (Map.Entry<String, Object> entry : jsonMap1.entrySet()) {
+            uuid = entry.getKey();
+            Object json1 = entry.getValue();
+            Object json2 = jsonMap2.get(uuid);
+            if (json2 != null) {
+                Map<String, Object> diff = findJsonDifferences(castToStringObjectMap(json1), castToStringObjectMap(json2));
+                if (!diff.isEmpty()) {
+                    differences.put(uuid, diff);
+                    hasDifferences = true;
+                }
+            } else {
+                differences.put(uuid, metadataDiffOutputGenerator.createFieldDiff( json1, null,"removed"));
+                hasDifferences = true;
+            }
+        }
+
+        for (Map.Entry<String, Object> entry : jsonMap2.entrySet()) {
+            String uuid2 = entry.getKey();
+            if (!jsonMap1.containsKey(uuid2)) {
+                differences.put(uuid2, metadataDiffOutputGenerator.createFieldDiff(null, entry.getValue(), "added"));
+                hasDifferences = true;
+            }
+        }
+
+        if (!hasDifferences) {
+            differences.put(uuid, metadataDiffOutputGenerator.createFieldDiff(null, null, "noModification"));
+        }
+        return differences;
+    }
+
+    protected Map<String, Object> findJsonDifferences(Map<String, Object> json1, Map<String, Object> json2) {
+        Map<String, Object> differences = new LinkedHashMap<>();
+        if (json1 == null && json2 == null) {
+            return differences;
+        }
+
+        if (json1 == null) {
+            json2.forEach((key, value) -> differences.put(key, metadataDiffOutputGenerator.createFieldDiff(null, value, "added")));
+            return differences;
+        }
+
+        if (json2 == null) {
+            json1.forEach((key, value) -> differences.put(key, metadataDiffOutputGenerator.createFieldDiff(value, null, "removed")));
+            return differences;
+        }
+
+        for (Map.Entry<String, Object> entry : json1.entrySet()) {
+            String key = entry.getKey();
+            Object value1 = entry.getValue();
+            Object value2 = json2.get(key);
+
+            if (key.equals("id")) {
+                continue;
+            }
+            if (value2 == null) {
+                differences.put(key, metadataDiffOutputGenerator.createFieldDiff(value1, null, "removed"));
+            } else {
+                if (value1 instanceof Map && value2 instanceof Map) {
+                    Map<String, Object> subDiff = findJsonDifferences((Map<String, Object>) value1, (Map<String, Object>) value2);
+                    if (!subDiff.isEmpty()) {
+                        differences.put(key, metadataDiffOutputGenerator.createObjectDiff((Map<String, Object>) value1, (Map<String, Object>) value2, "modified"));
+                    }
+                } else if (value1 instanceof List && value2 instanceof List) {
+                    List<Map<String, Object>> listDiff = findArrayDifferences((List<Object>) value1, (List<Object>) value2);
+                    if (!listDiff.isEmpty()) {
+                        differences.put(key, metadataDiffOutputGenerator.createArrayDiff((List<Object>) value1, (List<Object>) value2, "modified"));
+                    }
+                } else if (!value1.equals(value2)) {
+                    differences.put(key, metadataDiffOutputGenerator.createFieldDiff(value1, value2, "modified"));
+                }
+            }
+        }
+
+        for (Map.Entry<String, Object> entry : json2.entrySet()) {
+            String key = entry.getKey();
+            if (!json1.containsKey(key)) {
+                differences.put(key, metadataDiffOutputGenerator.createFieldDiff(null, entry.getValue(), "added"));
+            }
+        }
+
+        return differences;
+    }
+
+    protected List<Map<String, Object>> findArrayDifferences(List<Object> array1, List<Object> array2) {
+        List<Map<String, Object>> differences = new ArrayList<>();
+
+        Function<Map<String, Object>, String> getUuid = obj -> (String) obj.get("uuid");
+
+        Map<String, Map<String, Object>> map1 = array1.stream()
+                .filter(obj -> obj instanceof Map)
+                .map(obj -> (Map<String, Object>) obj)
+                .collect(Collectors.toMap(getUuid, Function.identity(), (e1, e2) -> e1));
+
+        Map<String, Map<String, Object>> map2 = array2.stream()
+                .filter(obj -> obj instanceof Map)
+                .map(obj -> (Map<String, Object>) obj)
+                .collect(Collectors.toMap(getUuid, Function.identity(), (e1, e2) -> e1));
+
+        for (String uuid : map2.keySet()) {
+            if (!map1.containsKey(uuid)) {
+                differences.add(metadataDiffOutputGenerator.createFieldDiff(null, map2.get(uuid), "added"));
+            } else {
+                Map<String, Object> obj1 = map1.get(uuid);
+                Map<String, Object> obj2 = map2.get(uuid);
+
+                Map<String, Object> subDiff = findJsonDifferences(obj1, obj2);
+                if (!subDiff.isEmpty()) {
+                    differences.add(metadataDiffOutputGenerator.createObjectDiff(obj1, obj2, "modified"));
+                }
+            }
+        }
+
+        for (String uuid : map1.keySet()) {
+            if (!map2.containsKey(uuid)) {
+                differences.add(metadataDiffOutputGenerator.createFieldDiff(map1.get(uuid), null, "removed"));
+            }
+        }
+        return differences;
+    }
+    private Map<String, Object> castToStringObjectMap(Object obj) {
+        if (obj instanceof Map) {
+            return (Map<String, Object>) obj;
+        }
+        return new HashMap<>();
+    }
+}
